LayoutTests/http/tests/webgpu/webgpu/shader/validation/resource_interface/bindings.html


                <script>
                    if (window.testRunner)
                        testRunner.waitUntilDone();
                    window.location.href = "/webgpu/cts.https.html?q=webgpu:shader,validation,resource_interface,bindings:*";
                </script>
            
