LayoutTests/http/tests/webgpu/webgpu/shader/validation/resource_interface/bindings-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Query `webgpu:shader,validation,resource_interface,bindings:*` does not match any cases

Harness Error (FAIL), message = Unhandled rejection: Query `webgpu:shader,validation,resource_interface,bindings:*` does not match any cases
